+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.19 Oct 25 2020
+- Add PyTorch Lightning logger (gorarakelyan)
+- Add TensorFlow v1 and v2 keras callbacks support (gorarakelyan)
+
 ## 2.0.18 Oct 7 2020
 - Add ability to run Aim UI in detached mode (gorarakelyan)
 - Add ability to specify repo path when running Aim UI (gorarakelyan)